requirements.txt
datasets>=2.19.0
faiss-cpu>=1.8.0
openai>=1.40.0
tiktoken>=0.7.0
rank-bm25>=0.2.2
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
tqdm>=4.66.0
python-dotenv>=1.0.1
pyyaml>=6.0
spacy>=3.7.0
# After installing spaCy, run: python -m spacy download en_core_web_sm
